Check overflow

+++ src/lib/fh_i.h
@@ -55,15 +55,15 @@ struct fh_parser;
 struct fh_compiler;
 struct fh_program;
 
-int fh_utf8_len(char *str, int str_size);
+int fh_utf8_len(char *str, size_t str_size);
 struct fh_src_loc fh_make_src_loc(uint16_t line, uint16_t col);
 void fh_output(struct fh_output *out, char *fmt, ...) __attribute__((format (printf, 2, 3)));
 void fh_dump_value(const struct fh_value *val);
 
 void fh_init_buffer(struct fh_buffer *buf);
 void fh_free_buffer(struct fh_buffer *buf);
-int fh_buf_grow(struct fh_buffer *buf, int add_size);
-int fh_buf_add_string(struct fh_buffer *buf, const char *str, int str_size);
+int fh_buf_grow(struct fh_buffer *buf, size_t add_size);
+int fh_buf_add_string(struct fh_buffer *buf, const char *str, size_t str_size);
 int fh_buf_add_byte(struct fh_buffer *buf, uint8_t c);
 
 struct fh_symtab *fh_new_symtab(void);
